# api/openapi.json
{
  "components": {
    "schemas": {
...
      "DecimalString": {
        "description": "An exact non-negative integer wire value that never depends on JavaScript\n`Number` precision.",
        "example": "42",
        "pattern": "^(0|[1-9][0-9]*)$",
        "type": "string"
      },
...
      "EmptyRequest": {
        "additionalProperties": false,
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Common secret-safe error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "ImageFormat": {
        "enum": [
          "png",
          "jpeg"
        ],
        "type": "string"
      },
...
      "Success_OperationResult": {
        "additionalProperties": false,
        "description": "Common successful JSON envelope.",
        "properties": {
          "data": {
            "additionalProperties": false,
            "properties": {
              "changed": {
                "type": "boolean"
              },
              "revision": {
                "$ref": "#/components/schemas/DecimalString"
              }
            },
            "required": [
              "changed",
              "revision"
            ],
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "Success_SavedScreenshot": {
        "additionalProperties": false,
        "description": "Common successful JSON envelope.",
        "properties": {
          "data": {
            "additionalProperties": false,
            "properties": {
              "display_path": {
                "type": "string"
              },
              "format": {
                "$ref": "#/components/schemas/ImageFormat"
              }
            },
            "required": [
              "display_path",
              "format"
            ],
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Closed REST and WebSocket contracts for PokeCon",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "PokeCon API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/camera/retry": {
      "post": {
        "operationId": "retry_camera",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_OperationResult"
                }
              }
            },
            "description": "Camera retry applied or a no-op"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Camera retry failure"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Camera backend failure"
          }
        },
        "tags": [
          "devices"
        ]
      }
    },
    "/api/camera/screenshot": {
      "post": {
        "operationId": "screenshot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_SavedScreenshot"
                }
              },
              "image/jpeg": {
                "schema": {
                  "items": {
                    "format": "int32",
                    "minimum": 0,
                    "type": "integer"
                  },
                  "type": "array"
                }
              },
              "image/png": {
                "schema": {
                  "items": {
                    "format": "int32",
                    "minimum": 0,
                    "type": "integer"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Saved screenshot metadata or downloaded image bytes",
            "headers": {
              "Content-Disposition": {
                "description": "Present for attachment responses",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No frame or incompatible UI mode"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid crop, name, path, or format"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Screenshot failure"
          }
        },
        "tags": [
          "devices"
        ]
      }
    },
    "/api/commands/control": {
      "post": {
        "operationId": "control_command",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_OperationResult"
                }
              }
            },
            "description": "Command transition applied or a no-op"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Command identity was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Command state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid command request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Command backend failure"
          }
        },
        "tags": [
          "commands"
        ]
      }
    },
    "/api/commands/reload": {
      "post": {
        "operationId": "reload_commands",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_OperationResult"
                }
              }
            },
            "description": "Command generation reloaded"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Profile switch conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Worker or cache failure"
          }
        },
        "tags": [
          "commands"
        ]
      }
    },
    "/api/devices/cameras": {
      "get": {
        "operationId": "get_cameras",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_Vec_CameraDevice"
                }
              }
            },
            "description": "Fresh camera enumeration"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Camera enumeration failure"
          }
        },
        "tags": [
          "devices"
        ]
      }
    },
    "/api/devices/serial-ports": {
      "get": {
        "operationId": "get_serial_ports",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_Vec_SerialPort"
                }
              }
            },
            "description": "Fresh serial-port enumeration"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Serial enumeration failure"
          }
        },
        "tags": [
          "devices"
        ]
      }
    },
...
    "/api/state": {
      "get": {
        "operationId": "get_state",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_StateSnapshot"
                }
              }
            },
            "description": "Current runtime state"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "State snapshot failed"
          }
        },
        "tags": [
          "state"
        ]
      }
    },
...
  },
...
}